% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uisapi-package.R
\docType{package}
\name{uisapi-package}
\alias{uisapi}
\alias{uisapi-package}
\title{uisapi: Access the UNESCO Institute for Statistics API}
\description{
Retrieve data from the UNESCO Institute for Statistics (UIS) API \url{https://api.uis.unesco.org/api/public/documentation/}. UIS provides public access to more than 4,000 indicators focusing on education, science and technology, culture, and communication.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/tidy-intelligence/r-uisapi}
  \item \url{https://tidy-intelligence.github.io/r-uisapi/}
  \item Report bugs at \url{https://github.com/tidy-intelligence/r-uisapi/issues}
}

}
\author{
\strong{Maintainer}: Christoph Scheuch \email{christoph@tidy-intelligence.com} (\href{https://orcid.org/0009-0004-0423-6819}{ORCID}) [copyright holder]

}
\keyword{internal}
